added verbose flag and fixed language of content for single page usages

+++ bin/support/content-helper.js
@@ -1,7 +1,17 @@
 import fs from 'fs';
+import Logger from './logger.js';
+import CLIHelper from './cli-helper.js';
 
 class ContentFileHelper {
     static replaceInFile(filePath, searchValue, replaceValue) {
+        if (!fs.existsSync(filePath)) {
+            CLIHelper.isVerboseModeEnabled() &&
+                Logger.warn(
+                    `Source does not exist for replacement, skipping: ${filePath}`,
+                );
+            return;
+        }
+
         const fileContent = fs.readFileSync(filePath, 'utf-8');
         const updatedContent = fileContent.replace(searchValue, replaceValue);
         fs.writeFileSync(filePath, updatedContent);
